M1219 move IC select new attribute modal trigger to select classifier guesses dropdown
